// src/app/business-module/work-order/share/enum/work-order-page-type.enum.ts
export enum WorkOrderPageTypeEnum {
  // 新增
  add =  'add',
  // 编辑
  update =  'update',
  // checklist 巡检报告
  checkList = 'checkList',
  // 未完工巡检工单详情
  unfinishedView = 'unfinishedView',
  // 巡检任务详情
  taskView = 'taskView',
  // 编辑(重新)
  restUpdate = 'restUpdate',
  // 详情
  view = 'view',
  // 已完工
  finished = 'finished',
  // 未完工
  unfinished = 'unfinished',
  // 告警
  alarm = 'alarm',
  // 历史
  history = 'history',
  // 重新生成
  rebuild = 'rebuild',
}